.pub-cache/hosted/pub.dartlang.org/html-0.15.2/test/data/tree-construction/regression_tests.dat

#data
<button><p></button><button></button>
#info
Regression test for https://github.com/dart-lang/html/issues/122
#errors
Errors are currently untested: https://github.com/dart-lang/html/issues/127
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <p>
|     <button>
